Fixed code to allow properties to use scoped values using '/' characters in property values
