+++ src/main/java/org/verifyica/pipeliner/execution/support/Ipc.java
@@ -33,7 +33,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.TreeMap;
 import org.verifyica.pipeliner.common.ShutdownHook;
-import org.verifyica.pipeliner.model.PropertyName;
+import org.verifyica.pipeliner.model.Property;
 
 /** Class to implement Ipc */
 public class Ipc {
@@ -90,13 +90,13 @@ public class Ipc {
                 int equalIndex = line.indexOf('=');
                 if (equalIndex == -1) {
                     String key = line.trim();
-                    if (PropertyName.isInvalid(key)) {
+                    if (Property.isInvalid(key)) {
                         throw new IpcException(format("invalid capture property [%s]", key));
                     }
                     map.put(key, "");
                 } else {
                     String key = line.substring(0, equalIndex).trim();
-                    if (PropertyName.isInvalid(key)) {
+                    if (Property.isInvalid(key)) {
                         throw new IpcException(format("invalid capture property [%s]", key));
                     }
                     String value = line.substring(equalIndex + 1);
