+++ tests/all.yaml
@@ -13,10 +13,14 @@ pipeline:
           run: $PIPELINER tests/test-properties-4.yaml
         - name: tests/test-properties-5.yaml
           run: $PIPELINER tests/test-properties-5.yaml
-        - name: tests/test-properties-7.yaml
-          run: $PIPELINER tests/test-properties-7.yaml
         - name: tests/test-properties-6.yaml
           run: $PIPELINER tests/test-properties-6.yaml
+        - name: tests/test-properties-7.yaml
+          run: $PIPELINER tests/test-properties-7.yaml
+        - name: tests/test-properties-8.yaml
+          run: $PIPELINER tests/test-properties-8.yaml
+        - name: tests/test-properties-8.yaml
+          run: $PIPELINER tests/test-properties-9.yaml
         - name: tests/test-properties-mask-1.yaml
           run: $PIPELINER tests/test-properties-mask-1.yaml
         - name: tests/test-properties-mask-2.yaml
