+++ tests/test-properties-8.yaml
@@ -1,0 +1,26 @@
+pipeline:
+  name: hello-world-pipeline
+  id: hello-world-pipeline
+  with:
+    test.scripts.directory: $PIPELINER_HOME/tests/scripts
+    pipeline.property: pipeline-foo
+  jobs:
+    - name: hello-world-job
+      id: hello-world-job
+      with:
+        job.property: ${{ hello-world-pipeline/pipeline.property }}/job-foo
+      steps:
+        - name: hello-world-step
+          id: hello-world-step
+          with:
+            step.property: ${{ hello-world-job/job.property }}/step-foo
+            step.property.2: ${{ job.property }}/step-foo
+          run: |
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ pipeline.property }}" "pipeline-foo"
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ pipeline.property }}" "${{ hello-world-pipeline/pipeline.property }}"
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ job.property }}" "${{ hello-world-job/job.property }}"
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ job.property }}" "${{ hello-world-pipeline/hello-world-job/job.property }}"
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ step.property }}" "pipeline-foo/job-foo/step-foo"
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ step.property }}" "${{ hello-world-job/hello-world-step/step.property }}"
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ step.property }}" "${{ hello-world-pipeline/hello-world-job/hello-world-step/step.property }}"
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ step.property }}" "${{ step.property.2 }}"
