+++ tests/test-properties-9.yaml
@@ -1,0 +1,26 @@
+pipeline:
+  name: hello-world-pipeline
+  id: hello-world-pipeline
+  with:
+    test.scripts.directory: $PIPELINER_HOME/tests/scripts
+    test.property: pipeline test property
+  jobs:
+    - name: hello-world-job
+      id: hello-world-job
+      with:
+        test.property: job test property
+      steps:
+        - name: hello-world-step
+          id: hello-world-step
+          with:
+            test.property: step test property
+          run: |
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ test.property }}" "step test property"
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ hello-world-step.test.property }}" "step test property"
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ hello-world-step/test.property }}" "step test property"
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ hello-world-job.test.property }}" "job test property"            
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ hello-world-job/test.property }}" "job test property"
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ hello-world-pipeline.hello-world-job.test.property }}" "job test property"            
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ hello-world-pipeline/hello-world-job/test.property }}" "job test property"
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ hello-world-pipeline.test.property }}" "pipeline test property"            
+            ${{test.scripts.directory}}/test-arguments-are-equal.sh "${{ hello-world-pipeline/test.property }}" "pipeline test property"
